# specs/001-matrades-product-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Matrades Application API
  version: 1.1.0
  description: >-
    Contract for authenticated configuration, research, strategy, risk, approval,
    reconciliation, monitoring, journal, and performance workflows. Broker execution
    writes are intentionally absent from V1.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
  /auth/signup:
    post:
      tags: [Authentication]
      security: []
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '202':
          description: Registration accepted; email verification required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSecurityState' }
        default: { $ref: '#/components/responses/Error' }
  /auth/sessions:
    post:
      tags: [Authentication]
      security: []
      operationId: beginSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, writeOnly: true }
      responses:
        '202':
          description: Primary authentication accepted; MFA challenge required
          content:
            application/json:
              schema:
                type: object
                required: [challenge_id, expires_at]
                properties:
                  challenge_id: { type: string, format: uuid }
                  expires_at: { type: string, format: date-time }
        default: { $ref: '#/components/responses/Error' }
    delete:
      tags: [Authentication]
      operationId: revokeCurrentSession
      responses:
        '204': { description: Session revoked }
        default: { $ref: '#/components/responses/Error' }
  /auth/email-verifications:
    post:
      tags: [Authentication]
      security: []
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        '200':
          description: Email verified; MFA enrollment required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSecurityState' }
        default: { $ref: '#/components/responses/Error' }
  /auth/mfa/enrollments:
    post:
      tags: [Authentication]
      operationId: createMfaEnrollment
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Pending enrollment and one-time setup material
          content:
            application/json:
              schema:
                type: object
                required: [enrollment_id, method, setup_uri]
                properties:
                  enrollment_id: { type: string, format: uuid }
                  method: { const: TOTP }
                  setup_uri: { type: string, writeOnly: true }
        default: { $ref: '#/components/responses/Error' }
  /auth/mfa/challenges:
    post:
      tags: [Authentication]
      operationId: verifyMfaChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge_id, code]
              properties:
                challenge_id: { type: string, format: uuid }
                code: { type: string, minLength: 6, maxLength: 12, writeOnly: true }
      responses:
        '200':
          description: MFA verified
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSecurityState' }
        default: { $ref: '#/components/responses/Error' }
  /auth/step-up:
    post:
      tags: [Authentication]
      operationId: verifyStepUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action_scope, code]
              properties:
                action_scope: { type: string }
                code: { type: string, writeOnly: true }
      responses:
        '200':
          description: Single-use scoped step-up grant
          content:
            application/json:
              schema:
                type: object
                required: [grant_id, action_scope, expires_at]
                properties:
                  grant_id: { type: string, format: uuid }
                  action_scope: { type: string }
                  expires_at: { type: string, format: date-time }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: matrades_session
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
...
  responses:
    Error:
      description: Structured safe error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, correlation_id]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        correlation_id: { type: string, format: uuid }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
    SignupRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12, writeOnly: true }
    UserSecurityState:
      type: object
      required: [user_id, status, next_action]
      properties:
        user_id: { type: string, format: uuid }
        status: { enum: [REGISTERED, EMAIL_VERIFIED, MFA_CONFIGURED, ACTIVE, LOCKED, DISABLED] }
        next_action: { enum: [VERIFY_EMAIL, ENROLL_MFA, COMPLETE_MFA, NONE] }
        recovery_codes:
          type: array
          writeOnly: true
          items: { type: string }
...